Default eviction policy depends on node tier
When cache preference is disabled, the default policy is used.

When cache preference is enabled, the pinned window policy is used
on search nodes and the default policy on other nodes.
