+++ x-pack/plugin/stateless/src/test/java/org/elasticsearch/xpack/stateless/cache/PinnedWindowEvictionPolicyTests.java
@@ -8,7 +8,6 @@
 package org.elasticsearch.xpack.stateless.cache;
 
 import org.elasticsearch.blobcache.shared.CacheRegion;
-import org.elasticsearch.blobcache.shared.EvictionPolicy;
 import org.elasticsearch.common.settings.ClusterSettings;
 import org.elasticsearch.common.settings.Setting;
 import org.elasticsearch.common.settings.Settings;
@@ -25,7 +24,6 @@ import java.util.Set;
 import static org.elasticsearch.blobcache.shared.SharedBlobCacheService.UNKNOWN_TIMESTAMP;
 import static org.elasticsearch.xpack.stateless.cache.PinnedWindowEvictionPolicy.PINNED_WINDOW_DURATION_SETTING;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.instanceOf;
 
 public class PinnedWindowEvictionPolicyTests extends ESTestCase {
 
@@ -45,18 +43,6 @@ public class PinnedWindowEvictionPolicyTests extends ESTestCase {
 
             clusterSettings.applySettings(Settings.builder().put(PINNED_WINDOW_DURATION_SETTING.getKey(), "12h").build());
             assertThat(policy.getPinnedWindowDuration(), equalTo(TimeValue.timeValueHours(12)));
-        }
-    }
-
-    public void testCreateEvictionPolicyReturnsPinnedWindowPolicyWhenBoostEnabled() {
-        final DeterministicTaskQueue taskQueue = new DeterministicTaskQueue();
-        final ClusterSettings clusterSettings = createClusterSettings(Settings.EMPTY);
-        final Settings settings = Settings.builder()
-            .put(StatelessSharedBlobCacheService.STATELESS_CACHE_BOOST_PREFERENCE_ENABLED_SETTING.getKey(), true)
-            .build();
-        try (var clusterService = ClusterServiceUtils.createClusterService(taskQueue.getThreadPool(), clusterSettings)) {
-            EvictionPolicy<FileCacheKey> policy = StatelessSharedBlobCacheService.createEvictionPolicy(settings, clusterService);
-            assertThat(policy, instanceOf(PinnedWindowEvictionPolicy.class));
         }
     }
